apps/meminfo/requirements.txt
groq>=0.4.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
